examples: move "args" to its own header
